Moved the code to pkg/github and into wrapper function GetRepositories
