+++ go.mod
@@ -35,6 +35,7 @@ require (
 	github.com/peterbourgon/diskv v2.0.1+incompatible // indirect
 	github.com/pierrec/lz4 v2.0.5+incompatible // indirect
 	github.com/prometheus/client_golang v0.9.3-0.20190127221311-3c4408c8b829
+	github.com/prometheus/common v0.2.0
 	github.com/rcrowley/go-metrics v0.0.0-20181016184325-3113b8401b8a // indirect
 	github.com/shurcooL/githubv4 v0.0.0-20190625031733-ee671ab25ff0
 	github.com/shurcooL/graphql v0.0.0-20181231061246-d48a9a75455f // indirect
